backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
sqlalchemy
alembic
pytest
pytest-asyncio
httpx
